+++ src/commands/config.js
@@ -1,6 +1,7 @@
 const { setSetting } = require('../helpers/dbModel');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { Permissions } = require('discord.js');
+const { setPerms, adminCommands, manageEmojisCommands } = require('../helpers/utilities');
 
 module.exports = {
 	data: new SlashCommandBuilder()
@@ -27,6 +28,45 @@ module.exports = {
 				.addBooleanOption(option =>
 					option.setName('flag')
 						.setDescription('Set flag')
+						.setRequired(true)))
+		.addSubcommand(subcommand =>
+			subcommand.setName('allownsfw')
+				.setDescription('Allows NSFW images in emoji.gg commands. Default: false')
+				.addBooleanOption(option =>
+					option.setName('flag')
+						.setDescription('Set flag')
+						.setRequired(true)))
+		.addSubcommand(subcommand =>
+			subcommand.setName('togglecommand')
+				.setDescription('Disables/Enables a specified command')
+				.addStringOption(option =>
+					option.setName('commandname')
+						.setDescription('Set flag')
+						.setRequired(true)
+						.addChoices([
+							['backupemojis', 'backupemojis'],
+							['botinfo', 'botinfo'],
+							['clapify', 'clapify'],
+							['copysteal', 'copysteal'],
+							['dancify', 'dancify'],
+							['displaystats', 'displaystats'],
+							['emoji', 'emoji'],
+							['enlargeemoji', 'enlargeemoji'],
+							['getcount', 'getcount'],
+							['leaderboard', 'leaderboard'],
+							['listemojis', 'listemojis'],
+							['packsearch', 'packsearch'],
+							['random', 'random'],
+							['removeunused', 'removeunused'],
+							['renameemoji', 'renameemoji'],
+							['resetdb', 'resetdb'],
+							['search', 'search'],
+							['stickerfy', 'stickerfy'],
+							['uploademoji', 'uploademoji'],
+						]))
+				.addBooleanOption(option =>
+					option.setName('flag')
+						.setDescription('Set flag')
 						.setRequired(true))),
 	async execute(interaction) {
 		if (!interaction.member.permissions.has(Permissions.FLAGS.ADMINISTRATOR)) {
@@ -36,11 +76,39 @@ module.exports = {
 			});
 		}
 
+		await interaction.deferReply({ ephemeral: true });
+
 		const setting = interaction.options.getSubcommand();
 		const flag = interaction.options.getBoolean('flag') ? 1 : 0;
 
-		setSetting(interaction.guild.id, setting, flag);
+		if (setting === 'togglecommand') {
+			const commandName = interaction.options.getString('commandname');
 
-		return interaction.reply({ content: `Setting \`${setting}\` set to \`${Boolean(flag)}\`.`, ephemeral: true });
+			if (adminCommands.includes(commandName)) {
+				// Sets admin commands...
+				interaction.guild.roles.cache.filter(role => role.permissions.has(Permissions.FLAGS.ADMINISTRATOR) && !role.managed).each(adminRole => {
+					setPerms(adminRole, adminCommands, flag);
+				});
+			}
+			else if (manageEmojisCommands.includes(commandName)) {
+				// Sets manage emojis commands...
+				interaction.guild.roles.cache.filter(role => role.permissions.has(Permissions.FLAGS.MANAGE_EMOJIS_AND_STICKERS) && !role.managed).each(manageEmojisRole => {
+					setPerms(manageEmojisRole, manageEmojisCommands, flag);
+				});
+			}
+			else {
+				// Sets @everyone commands...
+				const role = interaction.guild.roles.cache.get(interaction.guildId);
+				setPerms(role, [commandName], flag);
+			}
+
+			return interaction.editReply({ content: `\`/${commandName}\` ${flag ? 'enabled' : 'disabled'}.` });
+		}
+		else {
+			// Set config flag in serversettings db table
+			setSetting(interaction.guild.id, setting, flag);
+			return interaction.editReply({ content: `Setting \`${setting}\` set to \`${Boolean(flag)}\`.` });
+		}
+
 	},
 };